Mstop/requirements.txt
termcolor
pwinput
prettytable
mysql-connector-python
datetime
